+++ data_check/data_check.py
@@ -281,25 +281,25 @@ class DataCheck:
             print(dc_result.message)
         return results
 
-    def gen_expectation(self, sql_file: Path):
+    def gen_expectation(self, sql_file: Path, force=False):
         """
         Executes the query for a data_check test
         and stores the result in the csv file.
         """
         expect_result = self.get_expect_file(sql_file)
-        if not expect_result.exists():
+        if not expect_result.exists() or force:
             result = self.run_query(sql_file.read_text(encoding="UTF-8"))
             result.to_csv(expect_result, index=False)
             print(f"expectation written to {expect_result}")
         else:
             print(f"expectation skipped for {expect_result}")
 
-    def generate_expectations(self, files: List[Path]):
+    def generate_expectations(self, files: List[Path], force=False):
         """
         Generated a expected results file for each file if it doesn't exists yet.
         """
         for sql_file in self.expand_files(files):
-            self.gen_expectation(sql_file)
+            self.gen_expectation(sql_file, force)
 
     def test_connection(self) -> bool:
         """